Display the last tick on FixedScaleAxis
When using bar graphs with Chartist.FixedScaleAxis and a specified
divisor, the last tick is not displayed. E.g:

With divisor = 12 and max(series) = 96, the calculated graph ticks are
[0, 8, 16, 24, 32, 40, 48, 56, 64, 72, 80, 88]

There are indeed 12 ticks, but arguably we don't want to count 0 as a
tick. With the proposed change it would also display 96 at the end of
the axis:

[0, 8, 16, 24, 32, 40, 48, 56, 64, 72, 80, 88, 96]